+++ types/ramda/test/tap-tests.ts
@@ -1,0 +1,6 @@
+import * as R from 'ramda';
+
+() => {
+  const sayX = (x: number) => console.log('x is ' + x);
+  const a: number = R.tap(sayX, 100); // => 100
+};
