+++ types/ramda/test/uniq-tests.ts
@@ -1,0 +1,7 @@
+import * as R from 'ramda';
+
+() => {
+  R.uniq([1, 1, 2, 1]); // => [1, 2]
+  R.uniq([{}, {}]); // => [{}, {}]
+  R.uniq([1, '1']); // => [1, '1']
+};
